requirements.txt
requests>=2.31.0
flask>=3.0.0
waitress>=3.0.0
APScheduler>=3.10.0
PyYAML>=6.0
pycryptodome>=3.20.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
